python-backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
PyPDF2>=3.0.1
google-generativeai>=0.8.0
langgraph>=0.2.55
langchain>=0.3.0
langchain-google-genai>=2.0.7
langchain-community>=0.3.0
langchain-core>=0.2.43,<0.4.0
requests>=2.31.0
beautifulsoup4>=4.12.3
python-dotenv>=1.0.1
pydantic>=2.10.4
aiofiles>=24.1.0
nltk>=3.9.1
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
typing-extensions>=4.8.0
